pyproject.toml
[project]
name = "events-graph-rag"
version = "0.0.1"
description = "Search events in the graph database using hybrid RAG"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-neo4j",
    "neo4j",
    "langchain",
    "langchain_openai",
    "langchain-community",
    "python-dotenv",
    "langchain_mistralai",
    "langchain_groq",
    "langchain_anthropic",
    "langchain_google_genai",
    "scipy",
    "click",
]
[project.optional-dependencies]
dev = ["ruff", "ipykernel"]

[project.scripts]
events-rag = "events_graph_rag.cli:cli"

[tool.setuptools.packages.find]
include = ["events_graph_rag", "evals"]
exclude = ["db_data"]
